# api/specs/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: allbright Elite - Flash Loan Arbitrage Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet/config:
    put:
      operationId: updateWalletConfig
      tags: [wallet]
      summary: Update wallet config (RPC, private key)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletConfigBody"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
...
components:
  schemas:
...
    WalletConfigBody:
      type: object
      properties:
        rpcEndpoint:
          type: string
        privateKey:
          type: string
        pimlicoApiKey:
          type: string
        sweepMode:
          type: string
          enum: [AUTO_SWEEP, MANUAL_CLAIM]
...
    OperationResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message