# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * allbright Elite - Flash Loan Arbitrage Engine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type EngineStatusMode =
  (typeof EngineStatusMode)[keyof typeof EngineStatusMode];

export const EngineStatusMode = {
  SHADOW: "SHADOW",
  LIVE: "LIVE",
  STOPPED: "STOPPED",
} as const;

export interface EngineStatus {
  running: boolean;
  mode: EngineStatusMode;
  /** Seconds engine has been running */
  uptime?: number;
  walletAddress?: string;
  gaslessMode: boolean;
  pimlicoEnabled: boolean;
  scannerActive: boolean;
}

export type StartEngineRequestMode =
  (typeof StartEngineRequestMode)[keyof typeof StartEngineRequestMode];

export const StartEngineRequestMode = {
  SHADOW: "SHADOW",
  LIVE: "LIVE",
} as const;

export interface StartEngineRequest {
  mode: StartEngineRequestMode;
  /** Flash loan size in ETH */
  flashLoanSize?: number;
  /** Minimum profit margin percentage */
  minMarginPct?: number;
}

export interface EngineActionResult {
  success: boolean;
  message: string;
  mode?: string;
}

export type TradeStatus = (typeof TradeStatus)[keyof typeof TradeStatus];

export const TradeStatus = {
  PENDING: "PENDING",
  EXECUTED: "EXECUTED",
  FAILED: "FAILED",
  REVERTED: "REVERTED",
} as const;

export interface Trade {
  id: string;
  timestamp: string;
  status: TradeStatus;
  tokenIn?: string;
  tokenOut?: string;
  amountIn?: string;
  profit: number;
  profitUsd?: number;
  bribePaid?: number;
  gasUsed?: number;
  txHash?: string;
  protocol?: string;
  latencyMs?: number;
  blockNumber?: number;
}

export interface TradeListResponse {
  trades: Trade[];
  total: number;
}

export interface TradeSummary {
  totalProfitEth: number;
  totalProfitUsd: number;
  totalTrades: number;
  successRate: number;
  avgProfitPerTrade?: number;
  sessionProfitEth: number;
  sessionProfitUsd: number;
  tradesPerHour: number;
  topProtocol?: string;
  totalBribesPaid?: number;
}

export type StreamEventType =
  (typeof StreamEventType)[keyof typeof StreamEventType];

export const StreamEventType = {
  SCANNING: "SCANNING",
  DETECTED: "DETECTED",
  BRIBED: "BRIBED",
  EXECUTED: "EXECUTED",
  FAILED: "FAILED",
  REVERTED: "REVERTED",
} as const;

export interface StreamEvent {
  id: string;
  timestamp: string;
  type: StreamEventType;
  message: string;
  profit?: number;
  txHash?: string;
  blockNumber?: number;
  protocol?: string;
}

export interface StreamEventsResponse {
  events: StreamEvent[];
}

export type WalletInfoSweepMode =
  (typeof WalletInfoSweepMode)[keyof typeof WalletInfoSweepMode];

export const WalletInfoSweepMode = {
  AUTO_SWEEP: "AUTO_SWEEP",
  MANUAL_CLAIM: "MANUAL_CLAIM",
} as const;

export interface WalletInfo {
  address?: string;
  balanceEth?: number;
  balanceUsd?: number;
  sweepMode: WalletInfoSweepMode;
  /** Masked API key */
  pimlicoApiKey?: string;
  /** Masked RPC endpoint */
  rpcEndpoint?: string;
  gaslessEnabled: boolean;
}

export type WalletConfigBodySweepMode =
  (typeof WalletConfigBodySweepMode)[keyof typeof WalletConfigBodySweepMode];

export const WalletConfigBodySweepMode = {
  AUTO_SWEEP: "AUTO_SWEEP",
  MANUAL_CLAIM: "MANUAL_CLAIM",
} as const;

export interface WalletConfigBody {
  rpcEndpoint?: string;
  privateKey?: string;
  pimlicoApiKey?: string;
  sweepMode?: WalletConfigBodySweepMode;
}

export interface ProfitPoint {
  time: string;
  eth: number;
  usd: number;
}

export interface TelemetryData {
  sessionProfitEth: number;
  sessionProfitUsd: number;
  tradesPerHour: number;
  p99LatencyUs: number;
  avgLatencyUs?: number;
  memoryUsageMb?: number;
  cpuPercent?: number;
  blocksScanned?: number;
  opportunitiesDetected?: number;
  opportunitiesExecuted?: number;
  uptimeSeconds?: number;
  profitHistory?: ProfitPoint[];
}

export interface SystemSettings {
  flashLoanSizeEth: number;
  minMarginPct: number;
  maxBribePct: number;
  simulationMode: boolean;
  targetProtocols?: string[];
  maxSlippagePct?: number;
  /** Masked key */
  openaiApiKey?: string;
  /** Masked key */
  pimlicoApiKey?: string;
  visibleKpis?: string[];
}

export interface UpdateSettingsRequest {
  flashLoanSizeEth?: number;
  minMarginPct?: number;
  maxBribePct?: number;
  simulationMode?: boolean;
  targetProtocols?: string[];
  maxSlippagePct?: number;
  openaiApiKey?: string;
  pimlicoApiKey?: string;
  visibleKpis?: string[];
}

export interface OperationResult {
  success: boolean;
  message: string;
}

export type ListTradesParams = {
  limit?: number;
  status?: ListTradesStatus;
};

export type ListTradesStatus =
  (typeof ListTradesStatus)[keyof typeof ListTradesStatus];

export const ListTradesStatus = {
  PENDING: "PENDING",
  EXECUTED: "EXECUTED",
  FAILED: "FAILED",
  REVERTED: "REVERTED",
} as const;