# lib/api-zod/src/generated/types/walletConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * allbright Elite - Flash Loan Arbitrage Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletConfigBodySweepMode } from './walletConfigBodySweepMode';

export interface WalletConfigBody {
  rpcEndpoint?: string;
  privateKey?: string;
  pimlicoApiKey?: string;
  sweepMode?: WalletConfigBodySweepMode;
}